+++ classes/modes/picture-lazyload-front.php
@@ -123,8 +123,8 @@ class Picture_Lazyload_Front extends Mode implements Mode_Interface {
 			$default_img = $this->get_attachment_image_src( $this->size_or_img_name, 'thumbnail' );
 		}
 
-		if ( is_array( $default_img ) ) {
-			$main_content = str_replace( '%%default_img%%', reset( htmlspecialchars_decode( $default_img ) ), $main_content );
+		if ( ! empty( $default_img ) ) {
+			$main_content = str_replace( '%%default_img%%', htmlspecialchars_decode( $default_img ), $main_content );
 		}
 
 		// Add sources in main content tpl
@@ -134,10 +134,11 @@ class Picture_Lazyload_Front extends Mode implements Mode_Interface {
 		$classes = implode( ' ', $classes );
 
 		$attributes              = 'class="lazyload ' . esc_attr( $classes ) . '"';
+		$attributes              = $attributes . ' alt=""';
 		$content_with_attributes = str_replace( '%%attributes%%', $attributes, $content_with_sources );
 
 		// Add pixel on all
-		echo str_replace( '%%srcset%%', 'src="' . ARI_PIXEL . '"', $content_with_attributes );
+		echo str_replace( [ '%%srcset%%','%%srcgif%%' ], ['srcset="' . ARI_PIXEL . '"', 'src="' . ARI_PIXEL . '"'], $content_with_attributes );
 	}
 
 	/**